Auto-create hit_counter rows for new counts
When installing the extension on an existing wiki and/or when
viewing/creating new pages, it was not recording page views because
UPDATE is ignored if the condition (page_id) does not match any rows.

Use Database::upsert to look for the specified page_id which then
either insert the count=1 row, or perform the +1 operation.

Bug: T104081
Change-Id: I37abedb59cc3bfba2c601879da1495f46c69727b