Remove unnecessary HCUpdater::newForDB() function.
The signature of the DatabaseUpdater::newForDB() function has changed in MediaWiki 1.28. The HCUpdater class overrides this function. The signature of the overriding function either needs to be updated of the overriding function needs to be removed. Since the overriding function does not appear to be used, the latter seems to be the best approach, which will allow the class to retain backward compatibility with prior MediaWiki versions.

Change-Id: I2da3f2495aba28278df89c17bb4d177f0fb7bf6c