Use main cache to avoid issues with UTF-8 keys
The method `wfMemcKey` may use a different object cache class
than what is associated with `CACHE_ANYTHING`. Therefore the generated
cache key might not fit for the `CACHE_ANYTHING` cache.

For example, when `CACHE_MEMCACHED` is used, UTF-8 chars are not allowed
in the cache key.

Please cherry-pick this change to `REL1_31` also.

Change-Id: I2d6b0b2aadd211390eebf015fd2ca9a5a9374116